Update Attachment.lua
I had a bug were I was hitting the baseplate when I was nowhere lose to hitting it, along with the fact that my first set of "lines" after starting a hitbox never seem to register a hit. After I made the changes, those bugs dissappeared.

PS: This is my first time using GitHub.